refactor getStatus callback
